replaced the files so they fit docker
