+++ nginx/default.conf
@@ -1,0 +1,14 @@
+# nginx/nginx.conf
+upstream app_backhend{
+		server ${APP_1_HOST}:5000;
+		server ${APP_2_HOST}:5000;		
+	}
+	server{
+		listen 80;
+		location / {
+			proxy_pass http://app_backhend;
+			proxy_set_header Host $host;
+			proxy_set_header X-Real-IP $remote_addr;
+			proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+		}
+	}
