+++ requirements.txt
@@ -1,15 +1,8 @@
-blinker==1.9.0
-cffi==1.17.1
-click==8.1.8
-cryptography==44.0.0
-Flask==3.1.0
-Flask-SQLAlchemy==3.1.1
-greenlet==3.1.1
-itsdangerous==2.2.0
-Jinja2==3.1.5
-MarkupSafe==3.0.2
-pycparser==2.22
-PyMySQL==1.1.1
-SQLAlchemy==2.0.37
-typing_extensions==4.12.2
-Werkzeug==3.1.3
+Flask==2.1.1
+Flask-SQLAlchemy==2.5.1
+SQLAlchemy==1.4.41
+cryptography==39.0.1
+Jinja2==3.1.2
+MarkupSafe==2.1.5
+Werkzeug==2.2.3
+pymysql==1.0.2
